==> notebooks/00_Background.ipynb <==
# Harvest Tweet Corpus

## Overview

This notebook contains code to harvest Tweets that mention the following words and hashtags:

- `Bitcoin`
- `BTC`
- `#BTC`
- `#Bitcoin`

We'll use the [Twitter API for Academic Research][] for bulk access to Tweets, with the goal of building a corpus that spans eight years of Twitter history.

## Docs and other resources

- [Twitter API for Academic Research][] - overview of Academic-level access
- [Twitter Search API docs][] - Provides "premium"-level access for academic researchers to the *Full Archive Search*, which allows you to gather Tweets as far back as 2006.
- [Twitter Full Archive Search python example][]



[Twitter API for Academic Research]: https://developer.twitter.com/en/products/twitter-api/academic-research

[Twitter Search API docs]: https://developer.twitter.com/en/docs/twitter-api/premium/search-api/overview

[Twitter Full Archive Search python example]: https://github.com/twitterdev/Twitter-API-v2-sample-code/blob/main/Full-Archive-Search/full-archive-search.py


## Twitter Full Archive search

Below is a sample code based on the `Twitter Full Archive Search python example` (*see above for link to original code example*).

### Preliminary steps

The below code will only work if you use the standard `pipenv` workflow:

```
cd cryptocurrency1/
pipenv run jupyter lab
```

The above starts Jupyter Lab in the context of a virtual environment, which in turn makes the `TWITTER_BEARER_TOKEN` available to your code as an environment variable.

If you followed that procedure, the below code should work.

Step 1) Twitter Count and Comparison to bitcoin price by month per year -- show a correlation between discourse and price -- give examples that we see with december 2017 (skyrocket example) and major dip example 

Step 2) Twitter Harvesting -- for each month per year harvested tweets that met a specific quer (give query)
                           -- explain the reasons for the filterst that were added to the query 
                           -- explain that the original goal was to harvest 100,000 tweets for each month per year but we soon realized that 
                           for certain years, not enough tweets that met our query existed -- therefore we based this off of as much data as we 
                           collect and normalized it for comparison -- also provides insight into when the topic of bitcoin was
                           more popular than others 
                           --did this for each year -- and also collected the date it was created as well as reply, retweet, quote and like count 
                           
Step 3) Text Processing -- removed punctuation, emojis, cleaned text 

Step 4) TextBlob sentiment analysis 

Step 5) Vader Sentiment Intensity Analyzer 

Step 6) Calculate aggregate values 

Step 7) Visualization 

Further exploration: 
    --autoaggressive analysis to explore XYZ
    --explore how sentiment spread by taking into account replies, retweets, quotes and likes to determine virality of a particular sentiment 
    --see how this differs from fiat markets 
    --considering bitcoins main audience (techies), meme culture and media could play a large role in portraying sentiment -- deeper exploration
    into sentiment expressed through images and memes could provide a richer insight into what public sentiment can tell us about the bitcoin market
