==> notebooks/01_TweetCounter.ipynb <==
Purpose: Count the total number of tweets for each month per year using the query specified in the code below. (Will repeat multiple times for each year). Export into csv file. 

import requests
import os
import json
import csv
import dateutil.parser
import time 
import pandas as pd 
from datetime import datetime

# The below environment variable is in the .env file
# It becomes accessible when you run "pipenv run jupyter lab" to start
# this notebook from the command line.
bearer_token = os.environ.get("TWITTER_BEARER_TOKEN")

count_url = "https://api.twitter.com/2/tweets/counts/all"
                
                
def bearer_oauth(r):
    """
    Method required by bearer token authentication.
    """

    r.headers["Authorization"] = f"Bearer {bearer_token}"
    r.headers["User-Agent"] = "v2FullArchiveTweetCountsPython"
    return r


def connect_to_endpoint(url, params, next_token = None):
    params['next_token'] = next_token
    response = requests.request("GET", count_url, auth=bearer_oauth, params=params)
    if response.status_code != 200:
        raise Exception(response.status_code, response.text)
    return response.json()


final_counts_by_year = {}

final_results_json = {} 

#COUNT FOR 2014
start_list = ['2014-01-01T00:00:00Z', '2014-02-01T00:00:00Z', '2014-03-01T00:00:00Z', '2014-04-01T00:00:00Z', '2014-05-01T00:00:00Z', '2014-06-01T00:00:00Z', '2014-07-01T00:00:00Z', '2014-08-01T00:00:00Z', '2014-09-01T00:00:00Z', '2014-10-01T00:00:00Z', '2014-11-01T00:00:00Z', '2014-12-01T00:00:00Z']
end_list = ['2014-01-31T00:00:00Z', '2014-02-28T00:00:00Z', '2014-03-31T00:00:00Z', '2014-04-30T00:00:00Z', '2014-05-31T00:00:00Z', '2014-06-30T00:00:00Z', '2014-07-31T00:00:00Z', '2014-08-30T00:00:00Z', '2014-09-30T00:00:00Z', '2014-10-31T00:00:00Z', '2014-11-30T00:00:00Z', '2014-12-31T00:00:00Z']
month = ['1-2014', '2-2014', '3-2014', '4-2014', '5-2014', '6-2014', '7-2014', '8-2014', '9-2014', '10-2014', '11-2014', '12-2014']

for i in range(0, len(start_list)): 
    
    total_tweets = 0
    
    count = 0 
    max_count = 500
    flag = True
    next_token = None

    while flag:
            if count >= max_count: 
                break 
            print("--------------------------")
            print("Token: ", next_token)
            query_params = {'query':'(bitcoin OR #bitcoin) -is:nullcast',
                    'start_time': start_list[i],
                    'end_time': end_list[i],
                    'granularity':'day',
                    'next_token' : {}}
            json_response = connect_to_endpoint(count_url, query_params)
            result_count = json_response['meta']['total_tweet_count']

            if 'next_token' in json_response['meta']:

                # Save the token to use for next call

                next_token = json_response['meta']['next_token']

                print("Next Token: ", next_token)
                if result_count is not None and result_count > 0 and next_token is not None:
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")

            else:
                if result_count is not None and result_count > 0:
                    print("-------------------")
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")
                    time.sleep(5)
    
                flag = False
                next_token = None 

            final_counts_by_year[month[i]] = total_tweets

            time.sleep(5)
            print("Total number of results: ", total_tweets)  

#COUNT FOR 2015
start_list = ['2015-01-01T00:00:00Z', '2015-02-01T00:00:00Z', '2015-03-01T00:00:00Z', '2015-04-01T00:00:00Z', '2015-05-01T00:00:00Z', '2015-06-01T00:00:00Z', '2015-07-01T00:00:00Z', '2015-08-01T00:00:00Z', '2015-09-01T00:00:00Z', '2015-10-01T00:00:00Z', '2015-11-01T00:00:00Z', '2015-12-01T00:00:00Z']
end_list = ['2015-01-31T00:00:00Z', '2015-02-28T00:00:00Z', '2015-03-31T00:00:00Z', '2015-04-30T00:00:00Z', '2015-05-31T00:00:00Z', '2015-06-30T00:00:00Z', '2015-07-31T00:00:00Z', '2015-08-30T00:00:00Z', '2015-09-30T00:00:00Z', '2015-10-31T00:00:00Z', '2015-11-30T00:00:00Z', '2015-12-31T00:00:00Z']
month = ['1-2015', '2-2015', '3-2015', '4-2015', '5-2015', '6-2015', '7-2015', '8-2015', '9-2015', '10-2015', '11-2015', '12-2015']

for i in range(0, len(start_list)): 
    
    count = 0 
    max_count = 500
    flag = True
    next_token = None
    while flag:
            if count >= max_count: 
                break 
            print("--------------------------")
            print("Token: ", next_token)
            query_params = {'query':'(bitcoin OR #bitcoin) -is:nullcast ',
                    'start_time': start_list[i],
                    'end_time': end_list[i],
                    'granularity':'day',
                    'next_token' : {}}
            json_response = connect_to_endpoint(count_url, query_params)
            result_count = json_response['meta']['total_tweet_count']

            if 'next_token' in json_response['meta']:

                # Save the token to use for next call

                next_token = json_response['meta']['next_token']

                print("Next Token: ", next_token)
                if result_count is not None and result_count > 0 and next_token is not None:
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")

            else:
                if result_count is not None and result_count > 0:
                    print("-------------------")
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")
                    time.sleep(5)

                flag = False
                next_token = None 



            time.sleep(5)
            print("Total number of results: ", total_tweets)  
    
#COUNT FOR 2016
start_list = ['2016-01-01T00:00:00Z', '2016-02-01T00:00:00Z', '2016-03-01T00:00:00Z', '2016-04-01T00:00:00Z', '2016-05-01T00:00:00Z', '2016-06-01T00:00:00Z', '2016-07-01T00:00:00Z', '2016-08-01T00:00:00Z', '2016-09-01T00:00:00Z', '2016-10-01T00:00:00Z', '2016-11-01T00:00:00Z', '2016-12-01T00:00:00Z']
end_list = ['2016-01-31T00:00:00Z', '2016-02-28T00:00:00Z', '2016-03-31T00:00:00Z', '2016-04-30T00:00:00Z', '2016-05-31T00:00:00Z', '2016-06-30T00:00:00Z', '2016-07-31T00:00:00Z', '2016-08-30T00:00:00Z', '2016-09-30T00:00:00Z', '2016-10-31T00:00:00Z', '2016-11-30T00:00:00Z', '2016-12-31T00:00:00Z']
month = ['1-2016', '2-2016', '3-2016', '4-2016', '5-2016', '6-2016', '7-2016', '8-2016', '9-2016', '10-2016', '11-2016', '12-2016']

for i in range(0, len(start_list)): 
    
    count = 0 
    max_count = 500
    flag = True
    next_token = None
    while flag:
            if count >= max_count: 
                break 
            print("--------------------------")
            print("Token: ", next_token)
            query_params = {'query':'(bitcoin OR #bitcoin) -is:nullcast ',
                    'start_time': start_list[i],
                    'end_time': end_list[i],
                    'granularity':'day',
                    'next_token' : {}}
            json_response = connect_to_endpoint(count_url, query_params)
            result_count = json_response['meta']['total_tweet_count']

            if 'next_token' in json_response['meta']:

                # Save the token to use for next call

                next_token = json_response['meta']['next_token']

                print("Next Token: ", next_token)
                if result_count is not None and result_count > 0 and next_token is not None:
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")

            else:
                if result_count is not None and result_count > 0:
                    print("-------------------")
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")
                    time.sleep(5)

                flag = False
                next_token = None 



            time.sleep(5)
            print("Total number of results: ", total_tweets)  
    
#COUNT FOR 2017
start_list = ['2017-01-01T00:00:00Z', '2017-02-01T00:00:00Z', '2017-03-01T00:00:00Z', '2017-04-01T00:00:00Z', '2017-05-01T00:00:00Z', '2017-06-01T00:00:00Z', '2017-07-01T00:00:00Z', '2017-08-01T00:00:00Z', '2017-09-01T00:00:00Z', '2017-10-01T00:00:00Z', '2017-11-01T00:00:00Z', '2017-12-01T00:00:00Z']
end_list = ['2017-01-31T00:00:00Z', '2017-02-28T00:00:00Z', '2017-03-31T00:00:00Z', '2017-04-30T00:00:00Z', '2017-05-31T00:00:00Z', '2017-06-30T00:00:00Z', '2017-07-31T00:00:00Z', '2017-08-30T00:00:00Z', '2017-09-30T00:00:00Z', '2017-10-31T00:00:00Z', '2017-11-30T00:00:00Z', '2017-12-31T00:00:00Z']
month = ['1-2017', '2-2017', '3-2017', '4-2017', '5-2017', '6-2017', '7-2017', '8-2017', '9-2017', '10-2017', '11-2017', '12-2017']

for i in range(0, len(start_list)): 
    
    count = 0 
    max_count = 500
    flag = True
    next_token = None
    while flag:
            if count >= max_count: 
                break 
            print("--------------------------")
            print("Token: ", next_token)
            query_params = {'query':'(bitcoin OR #bitcoin) -is:nullcast ',
                    'start_time': start_list[i],
                    'end_time': end_list[i],
                    'granularity':'day',
                    'next_token' : {}}
            json_response = connect_to_endpoint(count_url, query_params)
            result_count = json_response['meta']['total_tweet_count']

            if 'next_token' in json_response['meta']:

                # Save the token to use for next call

                next_token = json_response['meta']['next_token']

                print("Next Token: ", next_token)
                if result_count is not None and result_count > 0 and next_token is not None:
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")

            else:
                if result_count is not None and result_count > 0:
                    print("-------------------")
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")
                    time.sleep(5)

                flag = False
                next_token = None 



            time.sleep(5)
            print("Total number of results: ", total_tweets)  

#COUNT FOR 2018
start_list = ['2018-01-01T00:00:00Z', '2018-02-01T00:00:00Z', '2018-03-01T00:00:00Z', '2018-04-01T00:00:00Z', '2018-05-01T00:00:00Z', '2018-06-01T00:00:00Z', '2018-07-01T00:00:00Z', '2018-08-01T00:00:00Z', '2018-09-01T00:00:00Z', '2018-10-01T00:00:00Z', '2018-11-01T00:00:00Z', '2018-12-01T00:00:00Z']
end_list = ['2018-01-31T00:00:00Z', '2018-02-28T00:00:00Z', '2018-03-31T00:00:00Z', '2018-04-30T00:00:00Z', '2018-05-31T00:00:00Z', '2018-06-30T00:00:00Z', '2018-07-31T00:00:00Z', '2018-08-30T00:00:00Z', '2018-09-30T00:00:00Z', '2018-10-31T00:00:00Z', '2018-11-30T00:00:00Z', '2018-12-31T00:00:00Z']
month = ['1-2018', '2-2018', '3-2018', '4-2018', '5-2018', '6-2018', '7-2018', '8-2018', '9-2018', '10-2018', '11-2018', '12-2018']

for i in range(0, len(start_list)): 
    
    count = 0 
    max_count = 500
    flag = True
    next_token = None
    while flag:
            if count >= max_count: 
                break 
            print("--------------------------")
            print("Token: ", next_token)
            query_params = {'query':'(bitcoin OR #bitcoin) -is:nullcast ',
                    'start_time': start_list[i],
                    'end_time': end_list[i],
                    'granularity':'day',
                    'next_token' : {}}
            json_response = connect_to_endpoint(count_url, query_params)
            result_count = json_response['meta']['total_tweet_count']

            if 'next_token' in json_response['meta']:

                # Save the token to use for next call

                next_token = json_response['meta']['next_token']

                print("Next Token: ", next_token)
                if result_count is not None and result_count > 0 and next_token is not None:
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")

            else:
                if result_count is not None and result_count > 0:
                    print("-------------------")
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")
                    time.sleep(5)

                flag = False
                next_token = None 



            time.sleep(5)
            print("Total number of results: ", total_tweets)  
    
    
#COUNT FOR 2019 
start_list = ['2019-01-01T00:00:00Z', '2019-02-01T00:00:00Z', '2019-03-01T00:00:00Z', '2019-04-01T00:00:00Z', '2019-05-01T00:00:00Z', '2019-06-01T00:00:00Z', '2019-07-01T00:00:00Z', '2019-08-01T00:00:00Z', '2019-09-01T00:00:00Z', '2019-10-01T00:00:00Z', '2019-11-01T00:00:00Z', '2019-12-01T00:00:00Z']
end_list = ['2019-01-31T00:00:00Z', '2019-02-28T00:00:00Z', '2019-03-31T00:00:00Z', '2019-04-30T00:00:00Z', '2019-05-31T00:00:00Z', '2019-06-30T00:00:00Z', '2019-07-31T00:00:00Z', '2019-08-30T00:00:00Z', '2019-09-30T00:00:00Z', '2019-10-31T00:00:00Z', '2019-11-30T00:00:00Z', '2019-12-31T00:00:00Z']
month = ['1-2019', '2-2019', '3-2019', '4-2019', '5-2019', '6-2019', '7-2019', '8-2019', '9-2019', '10-2019', '11-2019', '12-2019']

for i in range(0, len(start_list)): 
    
    count = 0 
    max_count = 500
    flag = True
    next_token = None
    while flag:
            if count >= max_count: 
                break 
            print("--------------------------")
            print("Token: ", next_token)
            query_params = {'query':'(bitcoin OR #bitcoin) -is:nullcast ',
                    'start_time': start_list[i],
                    'end_time': end_list[i],
                    'granularity':'day',
                    'next_token' : {}}
            json_response = connect_to_endpoint(count_url, query_params)
            result_count = json_response['meta']['total_tweet_count']

            if 'next_token' in json_response['meta']:

                # Save the token to use for next call

                next_token = json_response['meta']['next_token']

                print("Next Token: ", next_token)
                if result_count is not None and result_count > 0 and next_token is not None:
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")

            else:
                if result_count is not None and result_count > 0:
                    print("-------------------")
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")
                    time.sleep(5)

                flag = False
                next_token = None 



            time.sleep(5)
            print("Total number of results: ", total_tweets)  
        
#COUNT FOR 2019 
start_list = ['2019-01-01T00:00:00Z', '2019-02-01T00:00:00Z', '2019-03-01T00:00:00Z', '2019-04-01T00:00:00Z', '2019-05-01T00:00:00Z', '2019-06-01T00:00:00Z', '2019-07-01T00:00:00Z', '2019-08-01T00:00:00Z', '2019-09-01T00:00:00Z', '2019-10-01T00:00:00Z', '2019-11-01T00:00:00Z', '2019-12-01T00:00:00Z']
end_list = ['2019-01-31T00:00:00Z', '2019-02-28T00:00:00Z', '2019-03-31T00:00:00Z', '2019-04-30T00:00:00Z', '2019-05-31T00:00:00Z', '2019-06-30T00:00:00Z', '2019-07-31T00:00:00Z', '2019-08-30T00:00:00Z', '2019-09-30T00:00:00Z', '2019-10-31T00:00:00Z', '2019-11-30T00:00:00Z', '2019-12-31T00:00:00Z']
month = ['1-2019', '2-2019', '3-2021', '4-2019', '5-2019', '6-2019', '7-2019', '8-2019', '9-2019', '10-2019', '11-2019', '12-2019']

for i in range(0, len(start_list)): 
    
    count = 0 
    max_count = 500
    flag = True
    next_token = None
    while flag:
            if count >= max_count: 
                break 
            print("--------------------------")
            print("Token: ", next_token)
            query_params = {'query':'(bitcoin OR #bitcoin) -is:nullcast ',
                    'start_time': start_list[i],
                    'end_time': end_list[i],
                    'granularity':'day',
                    'next_token' : {}}
            json_response = connect_to_endpoint(count_url, query_params)
            result_count = json_response['meta']['total_tweet_count']

            if 'next_token' in json_response['meta']:

                # Save the token to use for next call

                next_token = json_response['meta']['next_token']

                print("Next Token: ", next_token)
                if result_count is not None and result_count > 0 and next_token is not None:
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")

            else:
                if result_count is not None and result_count > 0:
                    print("-------------------")
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")
                    time.sleep(5)

                flag = False
                next_token = None 



            time.sleep(5)
            print("Total number of results: ", total_tweets)  

#COUNT FOR 2020
start_list = ['2020-01-01T00:00:00Z', '2020-02-01T00:00:00Z', '2020-03-01T00:00:00Z', '2020-04-01T00:00:00Z', '2020-05-01T00:00:00Z', '2020-06-01T00:00:00Z', '2020-07-01T00:00:00Z', '2020-08-01T00:00:00Z', '2020-09-01T00:00:00Z', '2020-10-01T00:00:00Z', '2020-11-01T00:00:00Z', '2020-12-01T00:00:00Z']
end_list = ['2020-01-31T00:00:00Z', '2020-02-28T00:00:00Z', '2020-03-31T00:00:00Z', '2020-04-30T00:00:00Z', '2020-05-31T00:00:00Z', '2020-06-30T00:00:00Z', '2020-07-31T00:00:00Z', '2020-08-30T00:00:00Z', '2020-09-30T00:00:00Z', '2020-10-31T00:00:00Z', '2020-11-30T00:00:00Z', '2020-12-31T00:00:00Z']
month = ['1-2020', '2-2020', '3-2020', '4-2020', '5-2020', '6-2020', '7-2020', '8-2020', '9-2020', '10-2020', '11-2020', '12-2020']

for i in range(0, len(start_list)): 
    
    count = 0 
    max_count = 500
    flag = True
    next_token = None
    while flag:
            if count >= max_count: 
                break 
            print("--------------------------")
            print("Token: ", next_token)
            query_params = {'query':'(bitcoin OR #bitcoin) -is:nullcast ',
                    'start_time': start_list[i],
                    'end_time': end_list[i],
                    'granularity':'day',
                    'next_token' : {}}
            json_response = connect_to_endpoint(count_url, query_params)
            result_count = json_response['meta']['total_tweet_count']

            if 'next_token' in json_response['meta']:

                # Save the token to use for next call

                next_token = json_response['meta']['next_token']

                print("Next Token: ", next_token)
                if result_count is not None and result_count > 0 and next_token is not None:
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")

            else:
                if result_count is not None and result_count > 0:
                    print("-------------------")
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")
                    time.sleep(5)

                flag = False
                next_token = None 



            time.sleep(5)
            print("Total number of results: ", total_tweets)  

#COUNT FOR 2021
start_list = ['2021-01-01T00:00:00Z', '2021-02-01T00:00:00Z', '2021-03-01T00:00:00Z', '2021-04-01T00:00:00Z', '2021-05-01T00:00:00Z', '2021-06-01T00:00:00Z', '2021-07-01T00:00:00Z', '2021-08-01T00:00:00Z', '2021-09-01T00:00:00Z', '2021-10-01T00:00:00Z', '2021-11-01T00:00:00Z', '2021-12-01T00:00:00Z']
end_list = ['2021-01-31T00:00:00Z', '2021-02-28T00:00:00Z', '2021-03-31T00:00:00Z', '2021-04-30T00:00:00Z', '2021-05-31T00:00:00Z', '2021-06-30T00:00:00Z', '2021-07-31T00:00:00Z', '2021-08-30T00:00:00Z', '2021-09-30T00:00:00Z', '2021-10-31T00:00:00Z', '2021-11-30T00:00:00Z', '2021-12-31T00:00:00Z']
month = ['1-2021', '2-2021', '3-2021', '4-2021', '5-2021', '6-2021', '7-2021', '8-2021', '9-2021', '10-2021', '11-2021', '12-2021']

for i in range(0, len(start_list)): 
    
    count = 0 
    max_count = 500
    flag = True
    next_token = None
    while flag:
            if count >= max_count: 
                break 
            print("--------------------------")
            print("Token: ", next_token)
            query_params = {'query':'(bitcoin OR #bitcoin) -is:nullcast ',
                    'start_time': start_list[i],
                    'end_time': end_list[i],
                    'granularity':'day',
                    'next_token' : {}}
            json_response = connect_to_endpoint(count_url, query_params)
            result_count = json_response['meta']['total_tweet_count']

            if 'next_token' in json_response['meta']:

                # Save the token to use for next call

                next_token = json_response['meta']['next_token']

                print("Next Token: ", next_token)
                if result_count is not None and result_count > 0 and next_token is not None:
                    print("Start Date: ", start_list[i])
                    final_results_json[(month[i])] = json_response['meta']['total_tweet_count']
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")

            else:
                if result_count is not None and result_count > 0:
                    print("-------------------")
                    print("Start Date: ", start_list[i])
                    count += result_count
                    total_tweets += result_count
                    print("Total # of Tweets added: ", total_tweets)
                    print("-------------------")
                    time.sleep(5)

                flag = False
                next_token = None 



            time.sleep(5)
            print("Total number of results: ", total_tweets) 
            final_results_json[(month[i])] = json_response['meta']['total_tweet_count']

with open('Total_TweetCountFinal.json', 'w') as fp:
    json.dump(final_results_json, fp)

   

#Import Market Cap Data

data_dir = os.environ['DATA_DIR'] 
bitcoin_marketcap_csv = os.path.join(data_dir, 'BitcoinMarketCap.csv')
bitcoin_marketcap_df = pd.read_csv(bitcoin_marketcap_csv)
bitcoin_marketcap_df = bitcoin_marketcap_df.drop(bitcoin_marketcap_df.index[0:11])
bitcoin_marketcap_df['datetime'] = pd.to_datetime(bitcoin_marketcap_df['MonthYear'])
bitcoin_marketcap_df = bitcoin_marketcap_df.reset_index(drop=True)

bitcoin_marketcap_df

#Import Bitcoin price data 
bitcoin_price_csv = os.path.join(data_dir, 'bitcoinprice.csv')
bitcoin_price_df = pd.read_csv(bitcoin_price_csv)
bitcoin_price_df = bitcoin_price_df.dropna(axis=0, how="any", thresh=None, subset=None, inplace=False)
bitcoin_price_df['datetime'] = pd.to_datetime(bitcoin_price_df['MonthYear'], format='%b %Y')
bitcoin_price_df = bitcoin_price_df.reset_index(drop=True)

bitcoin_price_df

#Merge price and market cap data 
merged_df = pd.merge(bitcoin_price_df, bitcoin_marketcap_df[['market cap', 'datetime']], on='datetime' )
merged_df

with open ('Total_TweetCountFinal.json') as data: 
    count_dict = json.load(data)
    #dataframe = pd.DataFrame(count_dict)

df2 = pd.DataFrame.from_dict(count_dict, orient="index").reset_index().rename(columns={'index':'MonthYear', 0:'TweetCount'})
df2['datetime'] = pd.to_datetime(df2['MonthYear'], format='%m-%Y')

final_df = pd.merge(merged_df, df2[['TweetCount', 'datetime']], on='datetime')

final_df.to_csv('Preliminary_Count_Analysis.csv', sep = ',', index = False)
