==> notebooks/03_SentimentAnalysis.ipynb <==
Purpose: Use Vader sentiment analyzer to collect sentiment for each tweet it the dataframe. (Will repeat multiple times for each year)

import pandas as pd
from textblob import TextBlob
import re
import string
from nltk.sentiment.vader import SentimentIntensityAnalyzer
import time 

path = "/Users/cocoramgopal/code/cryptocurrency1/data/raw/"

#Example using 2014 data 
original_df = pd.read_csv(path+"2014Data.csv")
original_df.dtypes


#Drop unecessary columns and convert text to string 
original_df = original_df.drop(['retweet_count', 'reply_count', 'quote_count', 'like_count'], axis = 1)
original_df['text_cleaned_string'] = pd.Series(original_df['text'], dtype = pd.StringDtype())


#Convert Date to Datetime 
original_df['date_object'] = pd.to_datetime(original_df['created_at'], infer_datetime_format=True).dt.tz_localize(None)
original_df['datetime_striped'] = original_df['date_object'].dt.strftime('%b-%Y')
original_df['date_object'] =  original_df.date_object.dt.strftime("%Y-%m-%d") 

#Create new clean dataframe for sentiment analysis
clean_df = pd.DataFrame(original_df['text_cleaned_string'])
clean_df['month_year'] = original_df['datetime_striped']
clean_df['date'] =original_df['date_object']

#Function to process and clean tweets for analysis 

def process_tweets(tweet): 
    tweet = str(tweet)
    
    tweet = re.sub(r'\$\w*', '', tweet)

    # remove old style retweet text "RT"
    tweet = re.sub(r'^RT[\s]+', '', tweet)

	# remove hyperlinks
    tweet = re.sub(r'https?:\/\/.*[\r\n]*', '', tweet)
	
	# remove hashtags
	# only removing the hash # sign from the word
    tweet = re.sub(r'#', '', tweet)
    tweet = re.sub('r-', '', tweet)
    
    #remove html tags from tweets
    clean = re.compile('<.*?>')
    tweet = re.sub(clean, '', tweet)
    
    #remove mentions
    tweet = re.sub(r"@(\w+)", ' ', tweet, flags=re.MULTILINE)
    
    #remove /n 
    tweet=re.sub("\n","",tweet)
    
    #remove punctuation
    apostrophe_handled = re.sub("'", "'", tweet) 
    
    text  = re.sub("[^-9A-Za-z\s]", "" , apostrophe_handled)
    text = text.lower()
    return text

clean_df['text_cleaned_string'] = clean_df['text_cleaned_string'].apply(process_tweets)

#Function that applies the TextBlob sentiment analysis to each tweet 

def tb_enrich(ls):
    #Enriches a column of text with TextBlob Sentiment Analysis outputs
    tb_polarity = []
    tb_subject = []

    for tweet in ls:
        sentiment = TextBlob(tweet).sentiment
        tb_polarity.append(sentiment[0])
        tb_subject.append(sentiment[1])
    

    return tb_polarity, tb_subject

#Function called to determine overall sentiment 
def get_sentiment(ls):
    
    sentiment_list = []
    
    for tweet in ls: 
        if tweet > 0.00:
            sentiment = 'positive'
        elif tweet < 0.00:
            sentiment = 'negative'
        else:
            sentiment = 'neutral'
        sentiment_list.append(sentiment)
    return sentiment_list 

clean_df["Polarity"], clean_df["Subjectivity"] = tb_enrich(list(clean_df['text_cleaned_string']))
clean_df['polarity_sentiment'] = get_sentiment(list(clean_df['Polarity']))

clean_df['polarity_sentiment'].value_counts()

#Get sentiment scores from Vader Sentiment Analyzer 
clean_df['scores'] = clean_df['text_cleaned_string'].apply(lambda tweet: SentimentIntensityAnalyzer().polarity_scores(tweet))

#Create a new column for each of the returned scores -- positive, neutral, negative, compound 
clean_df['compound']  = clean_df['scores'].apply(lambda score_dict: score_dict['compound'])
clean_df['neutral']  = clean_df['scores'].apply(lambda score_dict: score_dict['neu'])
clean_df['negative']  = clean_df['scores'].apply(lambda score_dict: score_dict['neg'])
clean_df['positive']  = clean_df['scores'].apply(lambda score_dict: score_dict['pos'])

#Get overall sentiment (positive, negative or neutral)
clean_df['sentiment'] = get_sentiment(list(clean_df['compound']))

clean_df.to_csv('TestClean2014Data.csv', sep = ',', index = False)

#Compute average scores delivered by Vader for each month and combine into one dataframe 

scores_df = clean_df.groupby('month_year')['positive'].mean().reset_index(name = 'avg_positive')
scores_df2 = clean_df.groupby('month_year')['neutral'].mean().reset_index(name = 'avg_neutral')
scores_df3 = clean_df.groupby('month_year')['negative'].mean().reset_index(name = 'avg_negative')
scores_df4 = clean_df.groupby('month_year')['compound'].mean().reset_index(name = 'avg_compound')

merged_score1 = pd.merge(scores_df, scores_df2, on = 'month_year')
merged_score2 = merged_score1.merge(scores_df3, on = 'month_year')
merged_score3 = merged_score2.merge(scores_df4, on = 'month_year')

merged_score3['datetime'] = pd.to_datetime(merged_score3['month_year'])
merged_score3 = merged_score3.sort_values(by='datetime').reset_index(drop=True)

#Compute average polarity and subjectivity 
polarity_df = clean_df.groupby('month_year')['Polarity'].mean().reset_index(name = 'avg_polarity')
subjectivity_df = clean_df.groupby('month_year')['Subjectivity'].mean().reset_index(name = 'avg_subjectivity')

merged_polarity = pd.merge(polarity_df, subjectivity_df, on = 'month_year')
merged_polarity['datetime'] = pd.to_datetime(merged_polarity['month_year'])
merged_polarity = merged_polarity.sort_values(by='datetime')

#Merge all average scores together 
merged = pd.merge(merged_score3, merged_polarity[['month_year', 'avg_polarity']], on= 'month_year', how='left')
merged1 = pd.merge(merged, merged_polarity[['month_year', 'avg_subjectivity']], on= 'month_year', how='left')

merged1.to_csv('2017_avg_scores.csv', sep = ',', index = False)
