==> notebooks/04_MergeTables.ipynb <==
Purpose: Merge all of the tables that contain the sentiment analysis data for each year into one large dataframe. Export into csv file.  

# Notebook Setup:

import pandas as pd
from textblob import TextBlob
import re
import string
from nltk.sentiment.vader import SentimentIntensityAnalyzer
import time 

path = "/Users/cocoramgopal/code/cryptocurrency1/data/processed/"

# Import Clean Data For Each Year

df_2021 = pd.read_csv(path+'TestClean2021Data.csv')

df_2020 = pd.read_csv(path+'TestClean2020Data.csv')

df_2019 = pd.read_csv(path+'TestClean2019Data.csv')

df = pd.concat([df_2019, df_2020])

df2 = pd.concat([df, df_2021])

df_2018 = pd.read_csv(path+'TestClean2018Data.csv')
df3 = pd.concat([df_2018, df2])

df_2017 = pd.read_csv(path+'TestClean2017Data.csv')
df4 = pd.concat([df_2017, df3])

df_2016 = pd.read_csv(path+'TestClean2016Data.csv')
df5 = pd.concat([df_2016, df4])

df_2015 = pd.read_csv(path+'TestClean2015Data.csv')
df6 = pd.concat([df_2015, df5])

df_2014 = pd.read_csv(path+'TestClean2014Data.csv')
df7 = pd.concat([df_2014, df6])

df7.to_csv(path+'mer_daily_scores.csv', sep = ',', index = False)
